[UPDATE] Value tweaking

+++ Assets/Scripts/Runtime/GameModes/MainGameMode/MainGameModeConfig.cs
@@ -8,7 +8,7 @@ namespace GorillaGong.Runtime.GameModes.MainGameMode
     {
         [field: Header("Main Game Mode Specific")]
         [field: SerializeField]
-        public int DeathTime { get; private set; }
+        public float DeathTime { get; private set; }
         
         [field: SerializeField, Range(0f, 1f)] 
         public float DoubleInputProbability { get; private set; } = 0.25f;
